Add vehicle searching features

+++ db/migrate/20200501082150_create_model_features.rb
@@ -6,7 +6,7 @@ class CreateModelFeatures < ActiveRecord::Migration[5.2]
       t.string :series
       t.integer :body_type, index: true
       t.integer :cylinders
-      t.integer :displacement
+      t.float :displacement
       t.references :body_color, foreign_key: true
 
       t.timestamps
